+++ NEWS.md
@@ -1,6 +1,6 @@
 # v. 0.1.4
   
-  * 'h3_polyfill()` added; new dependencies on `sf`and `geojsonsf` have resulted.
+  * `h3_polyfill()` added; new dependencies on `sf`and `geojsonsf` have resulted.
 
 # v. 0.1.3